refact base kafka to reduce duplication of code
